+++ Transaction.h
@@ -9,20 +9,20 @@ class Transaction {
     int transactionId;
     int userId;
     int date;
-    double value;
+    float value;
     string kindOfTransaction;
 
 public:
     void setTransactionId(int newTransactionId);
     void setUserId(int newUserId);
     void setDate(int newDate);
-    void setValue(double newValue);
+    void setValue(float newValue);
     void setKindOfTransaction(string newKindOfTransaction);
 
     int getTransactionId();
     int getUserId();
     int getDate();
-    double getValue();
+    float getValue();
     string getKindOfTransaction();
 };
